Add file createList.js
