+++ src/index.html
@@ -14,7 +14,7 @@
         <i class="fa-solid fa-arrows-rotate"></i>
       </div>
       <div class="tdo-item">
-        <input type="text" placeholder="Add to your list..." />
+        <input id="add-to-List" type="text" placeholder="Add to your list..." />
       </div>
       <div id="to-do-list"></div>
       <div class="last-row"><p>Clear all completed</p></div>
